Moved MKL based classes to NativeWin64 project. Tests too.
