--- src/MGroup.LinearAlgebra.Implementations.NativeWin64/Class1.cs
@@ -1,9 +1,0 @@
-﻿namespace MGroup.LinearAlgebra.Implementations.NativeWin64
-{
-	using System;
-
-	public class Class1
-	{
-
-	}
-}
