+++ src/MGroup.LinearAlgebra/Implementations/Managed/LapackImplementations.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 //TODO: Some of these could be done by calling other BLAS, LAPACK functions. See the LAPACK source.
 //TODO: Use a port of reference LAPACK here. Custom implementations should be in another namespace (e.g. Factorizations or
