+++ tests/MGroup.LinearAlgebra.Tests/TestData/SquareInvertible10by10.cs
@@ -7,11 +7,11 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 	/// A square asymmetric 10-by-10 matrix that is invertible, but LU factorization needs pivoting.
 	/// Authors: Serafeim Bakalakos
 	/// </summary>
-	internal static class SquareInvertible10by10
+	public static class SquareInvertible10by10
 	{
-		internal const int Order = 10;
+		public const int Order = 10;
 
-		internal static double[,] Matrix => new double[,]
+		public static double[,] Matrix => new double[,]
 		{
 			{ 1.0000,    2.6667,    7.0000,    5.0000,    2.5000,    9.0000,    6.0000,    2.2500,    4.0000,    3.0000 },
 			{ 0.0000,    0.3333,    1.0000,   -2.5000,    5.5000,   -7.7500,    2.0000,    5.7500,    3.0000,    2.0000 },
@@ -25,19 +25,19 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ 3.0000,    7.0000,    9.0000,    4.0000,    2.6667,    9.0000,    8.0000,    5.0000,    7.0000,    3.0000 }
 		};
 
-		internal static double[] Lhs => new double[]
+		public static double[] Lhs => new double[]
 		{
 			3.4484, 1.9563, 2.7385, 4.2828, 5.3064, 4.3251, 0.1117, 4.0487, 2.6311, 2.6269
 		};
 
-		internal static double[] Rhs => MatrixOperations.MatrixTimesVector(Matrix, Lhs);
+		public static double[] Rhs => MatrixOperations.MatrixTimesVector(Matrix, Lhs);
 
-		internal static double[] Diagonal => new double[]
+		public static double[] Diagonal => new double[]
 		{
 			1, 0.3333, 4, 5, 0.25, 9, 8, 7, 7, 3
 		};
 
-		internal static double[,] FactorL => new double[,]
+		public static double[,] FactorL => new double[,]
 		{
 			{ 1.000000000000000, 0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000, 0.000000000000000},
 			{ 0.700000000000000, 1.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000,  0.000000000000000, 0.000000000000000},
@@ -51,7 +51,7 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ 0.066660000000000, 0.229167553739405,  0.540742649585241,  0.391641616972103,  0.581445473311093, -0.628021544221750,  0.073364914263990,  0.198333706840520, -0.659496414881728, 1.000000000000000}
 		};
 
-		internal static double[,] FactorU => new double[,]
+		public static double[,] FactorU => new double[,]
 		{
 			{ 5.000000000000000, 0.666700000000000, 4.500000000000000,  5.000000000000000, 0.250000000000000,  2.333300000000000, 1.500000000000000,  1.250000000000000,  9.000000000000000,  0.750000000000000 },
 			{ 0.000000000000000, 8.533310000000000, 5.850000000000000, -1.500000000000000, 7.825000000000000,  2.366690000000000, 1.450000000000000,  5.125000000000000,  0.699999999999999,  8.475000000000000 },
@@ -65,9 +65,9 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ 0.000000000000000, 0.000000000000000, 0.000000000000000,  0.000000000000000, 0.000000000000000,  0.000000000000000, 0.000000000000000,  0.000000000000000,  0.000000000000000,  3.279189768619025 }
 		};
 
-		internal const double Determinant = 17011962.9297146;
+		public const double Determinant = 17011962.9297146;
 
-		internal static double[,] Inverse => new double[,]
+		public static double[,] Inverse => new double[,]
 		{
 			{ -2.27411235611569928, -0.15826165906834855, -0.03971366478919014,  0.54754530193387962,  0.84564546633614235,  1.76721370456951843,  1.07391361538651520, -0.62155870054409057, -0.48250146850855319,  0.01019672027462176 },
 			{  0.22484785089916492,  0.00497907689343147,  0.13355139007048836, -0.21626167761339685, -0.27191124159200897, -0.36042272648326040, -0.27315652370363785,  0.13047889239071397,  0.11312595580665039,  0.24862184749402985 },
@@ -81,17 +81,17 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ -0.41339744413805679, -0.05827031189181564,  0.01456033811781599,  0.20111566009168913,  0.12490019370398427,  0.30495337890162211,  0.24042238135589683, -0.15636978776831514, -0.05449355077458019, -0.08532973391890820 }
 		};
 
-		internal static double[] EigenvaluesRealPart => new double[]
+		public static double[] EigenvaluesRealPart => new double[]
 		{
 			 39.1704534933583200, 9.6733808137362020, 6.8315684464554830, 4.9691564280223170, -1.4325846620495990, -1.4325846620495990, -5.1947609315368840, -5.1947609315368840, -0.7448864835033934, -2.0616815108959500
 		};
 
-		internal static double[] EigenvaluesImaginaryPart => new double[]
+		public static double[] EigenvaluesImaginaryPart => new double[]
 		{
 			 +0.0, +0.0, +0.0, +0.0, +5.408105558371837, -5.408105558371837, +0.7272146732164515, -0.7272146732164515, +0.0, +0.0
 		};
 
-		internal static double[,] EigenvectorsLeft => new double[,]
+		public static double[,] EigenvectorsLeft => new double[,]
 		{
 			{ +0.1346290250378348, +0.16150770481963270, -0.29513412929327440, +0.223337078514103100, -0.15141240255125860, -0.04812429314790514, -0.22114133726869400, -0.054132445578569330, -0.65982162942136910, +0.19208462373673850 },
 			{ +0.2627123928188636, +0.27432753000616160, +0.09564488225550430, +0.121663824652140400, +0.33807289333489820, +0.11629228650219570, +0.28287040653421700, -0.087989745384869950, -0.09604541959320620, +0.28832314771360460 },
@@ -105,7 +105,7 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ +0.3535843136529307, +0.05186988252484796, +0.03052283786235221, +0.005944164877591546, +0.18492249199097960, -0.37107576149820300, +0.65734341618436500, +0.000000000000000000, -0.11923930627148580, +0.54946796895876340 }
 		};
 
-		internal static double[,] EigenvectorsRight => new double[,]
+		public static double[,] EigenvectorsRight => new double[,]
 		{
 			{ +0.32678615875224180, -0.03263344036493219, -0.00430817666780283, -0.4908742294253686, -0.08260792908820513, -0.07502343378051754, -0.01316475674560672, -0.07668852381393969, -0.76811323270096440, -0.408953202901943400 },
 			{ +0.07060303833022699, +0.31729742995126800, +0.11750028290684330, +0.4535027518862313, -0.17797022347315430, -0.11703177803606610, -0.73666315551736770, -0000000000000000000, -0.01057525376492594, -0.388900537119509900 },
@@ -119,7 +119,7 @@ namespace MGroup.LinearAlgebra.Tests.TestData
 			{ +0.42210876958275950, +0.15330788612427290, +0.11978624511991890, -0.1520957828520478, -0.29015111137511290, +0.17737675992725380, +0.21571214416598800, -0.10128641322886940, -0.15869837683631010, -0.077252982667392540 }
 		};
 
-		internal static string FilePath => Directory.GetParent(Directory.GetCurrentDirectory()).Parent.Parent.FullName
+		public static string FilePath => Directory.GetParent(Directory.GetCurrentDirectory()).Parent.Parent.FullName
 			+ @"\MGroup.LinearAlgebra.Tests\Resources\SquareInvertible10by10.txt";
 	}
 }
